requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
arrow=1.2.2=pypi_0
asttokens=2.0.5=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2022.4.26=haa95532_0
certifi=2022.6.15=py39haa95532_0
cffi=1.15.0=py39h2bbff1b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
cmake=3.22.5=pypi_0
colorama=0.4.5=pypi_0
cpuonly=2.0=0
cryptography=37.0.1=py39h21b164f_0
cycler=0.11.0=pypi_0
cython=0.29.30=pypi_0
debugpy=1.6.2=pypi_0
decorator=5.1.1=pypi_0
distro=1.7.0=pypi_0
docker=5.0.3=pypi_0
docstring-parser=0.8.1=pypi_0
entrypoints=0.4=pypi_0
exceptiongroup=1.0.0rc8=pypi_0
executing=0.8.3=pypi_0
filelock=3.7.1=pypi_0
fonttools=4.34.4=pypi_0
freeplot=0.0.12=pypi_0
fsspec=2022.5.0=pypi_0
hypothesis=6.53.0=pypi_0
idna=3.3=pyhd3eb1b0_0
intel-openmp=2022.0.0=haa95532_3663
iopath=0.1.10=pypi_0
ipykernel=6.15.1=pypi_0
ipython=8.4.0=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.1.0=pypi_0
jupyter-client=7.3.4=pypi_0
jupyter-core=4.11.1=pypi_0
kiwisolver=1.4.4=pypi_0
libuv=1.40.0=he774522_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.2=pypi_0
matplotlib-inline=0.1.3=pypi_0
mkl=2022.0.0=haa95532_115
mypy-extensions=0.4.3=pypi_0
nest-asyncio=1.5.5=pypi_0
ninja=1.10.2.3=pypi_0
numpy=1.23.1=pypi_0
openssl=1.1.1q=h2bbff1b_0
packaging=21.3=pypi_0
pandas=1.4.3=pypi_0
parso=0.8.3=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.2.0=pypi_0
pip=22.1.2=py39haa95532_0
portalocker=2.3.0=py39haa95532_0
prompt-toolkit=3.0.30=pypi_0
psutil=5.9.1=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=8.0.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.2=pypi_0
pygments=2.12.0=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pyre-extensions=0.0.29=pypi_0
pysocks=1.7.1=py39haa95532_0
python=3.9.12=h6244533_0
python-dateutil=2.8.2=pypi_0
pytorch=1.12.0=py3.9_cpu_0
pytorch-mutex=1.0=cpu
pytz=2022.1=pypi_0
pywin32=227=pypi_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=pypi_0
requests=2.28.1=py39haa95532_0
scienceplots=1.0.9=pypi_0
scikit-build=0.15.0=pypi_0
scikit-learn=1.1.1=pypi_0
scipy=1.8.1=pypi_0
seaborn=0.11.2=pypi_0
setuptools=61.2.0=py39haa95532_0
six=1.16.0=pypi_0
sortedcontainers=2.4.0=pypi_0
sqlite=3.38.5=h2bbff1b_0
stack-data=0.3.0=pypi_0
tabulate=0.8.10=pypi_0
threadpoolctl=3.1.0=pypi_0
torchdata=0.4.0=py39
torchmetrics=0.9.2=pypi_0
torchrec=0.1.0=pypi_0
torchx-nightly=2022.7.24=pypi_0
tornado=6.2=pypi_0
tqdm=4.64.0=pypi_0
traitlets=5.3.0=pypi_0
typing=3.7.4.3=pypi_0
typing-inspect=0.7.1=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py39haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pypi_0
websocket-client=1.3.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39haa95532_2
